Pin GitHub Actions in workflows to commit SHA
Used [mheap/pin-github-action](https://github.com/mheap/pin-github-action) to
update the workfows in `.github/workflows/*.yml`

The utility updates the actions in the workflow to the specific SHA
commit referenced at that particular version. It leaves the specific
version pinned as a comment alongside the action.

This change follows best practices documented by:
* [GitHub](https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions)
* [OpenSSF](https://github.com/ossf/scorecard/blob/7206a2bdeb7020ab45744a1a6234c9ddf2f3713c/docs/checks.md#pinned-dependencies)